Add oslo-log as BR
oslo-config has added a optional requirement on oslo-log [1]. However
this is mandatory to run unit tests. An additional problem is that
oslo-log depends on oslo-config so this can lead to circular
dependencies. To allow repo bootstrap I'm adding a new parameter
repo_bootstrap that must be set to 1 to build oslo-config when
oslo-log does not exist in a repo.

[1] https://review.openstack.org/#/c/560094/

Change-Id: I15d0c926885aba76b57b9b649550d867f7f7f694