Move all doc-related BRs to the -doc subpackage
We had the python3-openstackdocstheme BR outside the -doc subpackage,
so let's remove it.

Also, when building without docs, we still had python2-sphinx in the
build requirements, since a unit test depended on it. To allow us
to remove it, we are removing that unit test when building without
docs.

Change-Id: If92ab31a6863711dfae2dbdb07cf48860e649e97